// apps/frontend2/src/api/models/createProjectRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectRequestUrl } from './createProjectRequestUrl';

export interface CreateProjectRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  url?: CreateProjectRequestUrl;
}