// apps/frontend2/src/api/models/profileProjectDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileProjectDTOUrl } from './profileProjectDTOUrl';

export interface ProfileProjectDTO {
  id: number;
  name: string;
  description: string;
  url: ProfileProjectDTOUrl;
  created_at: string;
  updated_at: string;
}