// apps/frontend2/src/api/models/updateProjectRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectRequestName } from './updateProjectRequestName';
import type { UpdateProjectRequestDescription } from './updateProjectRequestDescription';
import type { UpdateProjectRequestUrl } from './updateProjectRequestUrl';

export interface UpdateProjectRequest {
  name?: UpdateProjectRequestName;
  description?: UpdateProjectRequestDescription;
  url?: UpdateProjectRequestUrl;
}